feat: use mill for building, no more sbt
Also rename modules to shorter names.
Makefile and README for accessing common mill commands.

delete old modules
