+++ .gitignore
@@ -13,7 +13,12 @@ src_managed/
 project/boot/
 project/plugins/project/
 
+# mill
+out/
+
 # Scala-IDE specific
 .scala_dependencies
 .worksheet
 .idea
+.idea_modules
+*.iml
