--- fsm-cooking/README.md
@@ -1,18 +1,0 @@
-# Cooking FSM Demo
-
-This project is for the following blog posts:
-
-* Part 1: https://michaelzg.com/akka/2018/09/03/akka-actor-fsm-comparison-part1/
-* Part 2: To be published
-
-### Run
-
-```
-sbt "project fsm" run
-```
-
-### Test
-
-```
-sbt "project fsm" test
-```
