+++ fsm/README.md
@@ -1,0 +1,18 @@
+# Cooking FSM Demo
+
+This project is for the following blog posts:
+
+* Part 1: https://michaelzg.com/akka/2018/09/03/akka-actor-fsm-comparison-part1/
+* Part 2: To be published
+
+### Run 
+
+```
+mill fsm.run
+```
+
+### Test
+
+```
+mill.fsm.test
+```
