+++ streaming-merge/src/Tweet.scala
