+++ streaming-merge/src/TweetStore.scala
